fix register

+++ routes/api.php
@@ -2,6 +2,8 @@
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AuthController;
+use App\Http\Middleware\Authenticate;
+use App\Http\Middleware\Authorization;
 
 Route::group([
     'middleware' => 'api',
@@ -12,8 +14,8 @@ Route::group([
     Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:api')->name('logout');
     Route::post('/refresh', [AuthController::class, 'refresh'])->middleware('auth:api')->name('refresh');
     Route::post('/me', [AuthController::class, 'me'])->middleware('auth:api')->name('me');
-    Route::put('/{id}', [AuthController::class, 'update'])->middleware('auth:api')->name('update');
-    Route::delete('/{id}', [AuthController::class, 'delete'])->middleware('auth:api')->name('delete');
-    Route::post('/allUser', [AuthController::class, 'allUser'])->middleware('auth:api')->name('allUser');
+    Route::put('/{id}', [AuthController::class, 'update'])->middleware('auth:api')->name('update')->middleware(Authorization::class . ':admin');
+    Route::delete('/{id}', [AuthController::class, 'delete'])->middleware('auth:api')->name('delete')->middleware(Authorization::class . ':admin');
+    Route::post('/allUser', [AuthController::class, 'allUser'])->middleware('auth:api')->name('allUser')->middleware(Authorization::class . ':admin');
 });
